pyproject.toml
[project]
name = "winwin-cli"
version = "0.1.0"
description = "CLI 封装工具，专为 AI 使用设计"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "xuruikun"}
]
keywords = ["cli", "ai", "knowledge-base", "search", "document-conversion"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Utilities",
]
dependencies = [
    "click>=8.1.0",
    "pyyaml>=6.0.0",
    "orjson>=3.9.0",
    "jieba>=0.42.0",
    "rank-bm25>=0.2.0",
    "pydantic>=2.0.0",
    "tqdm>=4.65.0",
    "markitdown>=0.1.4",
]

[project.urls]
Homepage = "https://github.com/yourusername/winwin-cli"
Repository = "https://github.com/yourusername/winwin-cli"
Issues = "https://github.com/yourusername/winwin-cli/issues"

[project.scripts]
winwin-cli = "winwin_cli.cli:main"

[dependency-groups]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
]

[tool.uv]
package = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"